test: separate acceptance test wrappers into another folder
So that it can be used by another set of tests later
