+++ .gitignore
@@ -11,5 +11,6 @@ icon.png
 codecov.sh
 .nyc_output
 
+black-box-tests/out
 black-box-tests/test-environment/**/*
 !black-box-tests/test-environment/tsconfig.json
