+++ black-box-tests/post-compile/package.json
