+++ black-box-tests/shared/.eslintrc.js
@@ -1,0 +1,5 @@
+module.exports = {
+  parserOptions: {
+    project: `${__dirname}/tsconfig.json`
+  }
+}
