+++ black-box-tests/shared/generate-wrapper.ts
