+++ black-box-tests/shared/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "rootDir": ".",
+    "outDir": "../out",
+    "noEmit": false,
+    "composite": true
+  },
+  "include": ["**/*.ts"]
+}
